+++ deployer/models/__init__.py
@@ -1,0 +1,7 @@
+"""
+Data models for the Klik Token Deployer
+"""
+
+from .deployment import DeploymentRequest
+
+__all__ = ['DeploymentRequest']