+++ deployer/services/__init__.py
@@ -1,0 +1,7 @@
+"""
+Services for the Klik Token Deployer
+"""
+
+from .ipfs_service import IPFSService
+
+__all__ = ['IPFSService']